requirements.txt
pyTelegramBotAPI
python-dotenv
requests
Flask
gunicorn
waitress